Realizar entrada e saída de dados do usuário

+++ Modulo.02/atividade04.py
@@ -1,7 +1,7 @@
 # SOFTEX-RECIFE
 # Aluno: Fábio de Tássio
 
-# Atividade 02 do Módulo 04 (Executar função de códigos)
+# Atividade 04 do Módulo 04 (Executar função de códigos)
 
 # Faça uma função calculadora de dois números com três parâmetros: os dois primeiros serão os números da
 # operação e o terceiro será a entrada que definirá a operação a ser executada. Considera a seguinte definição:
@@ -32,6 +32,6 @@ def calculadora(num1, num2, oper):
 
 num1 = 3
 num2 = 7
-oper = float(input('Digite 1 - soma, 2 - subtração, 3 - multiplicação, 4 - divisão:'))
+oper = int(input('Digite 1 - soma, 2 - subtração, 3 - multiplicação, 4 - divisão:'))
 resultado = calculadora(num1,num2, oper)
 print(resultado)
